requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
cycler==0.10.0
Flask==1.1.2
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.14.1
kiwisolver==1.2.0
MarkupSafe==1.1.1
patsy==0.5.1
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
scipy==1.4.1
selenium==3.141.0
six==1.14.0
twilio==6.38.1
urllib3==1.25.9
Werkzeug==1.0.1
